Attention_Identifier/requirements.txt
opencv-python
numpy
requests
python-dotenv
ultralytics
